fix: enable and fix browser tests
