+++ package.json
@@ -131,8 +131,12 @@
     "lint": "aegir lint",
     "build": "aegir build",
     "generate": "protons ./src/message/dht.proto",
-    "test": "aegir test -t node",
-    "test:node": "aegir test -t node",
+    "test": "aegir test",
+    "test:node": "aegir test -t node -f './dist/test/**/*.{node,spec}.js' --cov",
+    "test:chrome": "aegir test -t browser -f './dist/test/**/*.spec.js' --cov",
+    "test:chrome-webworker": "aegir test -t webworker -f './dist/test/**/*.spec.js'",
+    "test:firefox": "aegir test -t browser -f './dist/test/**/*.spec.js' -- --browser firefox",
+    "test:firefox-webworker": "aegir test -t webworker -f './dist/test/**/*.spec.js' -- --browser firefox",
     "dep-check": "aegir dep-check",
     "release": "aegir release"
   },
@@ -159,6 +163,7 @@
     "any-signal": "^3.0.0",
     "datastore-core": "^8.0.1",
     "err-code": "^3.0.1",
+    "events": "^3.3.0",
     "hashlru": "^2.3.0",
     "interface-datastore": "^7.0.0",
     "it-all": "^1.0.6",
@@ -191,7 +196,7 @@
     "@types/lodash.range": "^3.2.6",
     "@types/varint": "^6.0.0",
     "@types/which": "^2.0.1",
-    "aegir": "^37.0.8",
+    "aegir": "^37.3.0",
     "datastore-level": "^9.0.0",
     "delay": "^5.0.0",
     "execa": "^6.0.0",
