+++ test/message.spec.ts
@@ -2,15 +2,11 @@
 
 import { expect } from 'aegir/chai'
 import { multiaddr } from '@multiformats/multiaddr'
-import range from 'lodash.range'
 import random from 'lodash.random'
 import { Libp2pRecord } from '@libp2p/record'
-import fs from 'fs'
-import path from 'path'
 import { Message, MESSAGE_TYPE } from '../src/message/index.js'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
-import { isPeerId } from '@libp2p/interface-peer-id'
 
 describe('Message', () => {
   it('create', () => {
@@ -87,26 +83,5 @@ describe('Message', () => {
 
     msg.clusterLevel = 10
     expect(msg.clusterLevel).to.eql(9)
-  })
-
-  it('go-interop', () => {
-    range(1, 9).forEach((i) => {
-      const raw = fs.readFileSync(
-        path.join(process.cwd(), 'test', 'fixtures', `msg-${i}`)
-      )
-
-      const msg = Message.deserialize(raw)
-
-      expect(msg.clusterLevel).to.gte(0)
-      if (msg.record != null) {
-        expect(msg.record.key).to.be.a('Uint8Array')
-      }
-
-      if (msg.providerPeers.length > 0) {
-        msg.providerPeers.forEach((p) => {
-          expect(isPeerId(p.id)).to.be.true()
-        })
-      }
-    })
   })
 })
